feat(server): improve project structure
