--- server/lib/controller.js
@@ -1,9 +1,0 @@
-const database = require("./database");
-
-let controller = {
-    getData: database.getData,
-    getDataById: database.getDataById,
-    saveData: database.saveData
-}
-
-module.exports = controller;
