+++ server/routes/data.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+const dataController = require('../controllers/data');
+
+router.get('/', dataController.getData);
+router.get('/:id', dataController.getDataById);
+router.post('/', dataController.saveData);
+
+module.exports = router;
