+++ server/server.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const app = express();
 const port = 3000;
-
-const controller = require("./lib/controller");
+const dataRouter = require("./routes/data");
 
 app.use(express.json());
 app.use((req, res, next) => {
@@ -10,25 +9,7 @@ app.use((req, res, next) => {
   next();
 });
 
-app.get("/data", (req, res) => {
-    let data = controller.getData();
-    console.log("Saved data until now:", data);
-    res.send(data);
-});
-
-app.get("/data/:id", (req, res) => {
-    let data = controller.getDataById(req.params.id);
-    res.send(data);
-});
-
-app.post("/", (req, res) => {
-    let pluginType = req.header("X-Plugin-Name");
-    let machineId = req.header("X-Machine-Id");
-    let msg = `Data from plugin '${pluginType}' received successfully.`;
-    console.log(msg);
-    controller.saveData(req.body, pluginType, machineId);
-    res.status(200).send(msg);
-});
+app.use('/data', dataRouter);
 
 app.listen(port, () => {
     console.log(`Server listening on port ${port}!`);
